[examples] update cors example
* use header_filter to add response headers
* properly detect CORS request
* handle both preflight and non-preflight requests
