+++ examples/cors/Dockerfile
@@ -1,5 +1,5 @@
 FROM registry.access.redhat.com/3scale-amp20/apicast-gateway:1.0
 
-COPY apicast_cors.lua /opt/app-root/app/src/
+COPY cors.lua /opt/app-root/app/src/
 COPY cors.conf /opt/app-root/app/apicast.d/
